Updates for new charts
Added figure editing capabilities for the new forest plot options.

Also, found and corrected some errors in the line edit creation widget (some of the objects were returned in reverse order so there was a mismatch between which object was which) and a piece of text that wasn't being pushed through the language fetcher.